feat: 입력한 포스트 정보 db에 추가 기능 구현

+++ src/main/java/com/pentoryall/post/dto/PostDTO.java
@@ -17,8 +17,8 @@ public class PostDTO {
     private String content;
     private String confirmContent;
     private String thumbnailImage;
-    private int views;
-    private int price;
+    private long views;
+    private long price;
     private char isPaid;
     private char isAdult;
     private char isPublic;
